Change the edit dialog to a panel modal
