+++ client/src/game/ui/selection/edit_dialog/ExtraSettings.vue
@@ -1,0 +1,119 @@
+<script lang="ts">
+import Vue from "vue";
+import Component from "vue-class-component";
+
+import { Prop } from "vue-property-decorator";
+
+import { Shape } from "@/game/shapes/shape";
+import { EventBus } from "@/game/event-bus";
+
+@Component
+export default class AccessSettings extends Vue {
+    @Prop() owned!: boolean;
+    @Prop() shape!: Shape;
+
+    updateAnnotation(event: { target: HTMLInputElement }): void {
+        if (!this.owned) return;
+        this.shape.setAnnotation(event.target.value, true);
+    }
+
+    openLabelManager(): void {
+        EventBus.$emit("LabelManager.Open");
+    }
+
+    removeLabel(uuid: string): void {
+        if (!this.owned) return;
+        this.shape.removeLabel(uuid, true);
+    }
+}
+</script>
+
+<template>
+    <div class="panel restore-panel">
+        <div class="spanrow header" v-t="'common.labels'"></div>
+        <div id="labels" class="spanrow">
+            <div v-for="label in shape.labels" class="label" :key="label.uuid">
+                <template v-if="label.category">
+                    <div class="label-user">{{ label.category }}</div>
+                    <div class="label-main" @click="removeLabel(label.uuid)">{{ label.name }}</div>
+                </template>
+                <template v-if="!label.category">
+                    <div class="label-main" @click="removeLabel(label.uuid)">{{ label.name }}</div>
+                </template>
+            </div>
+            <div class="label" id="label-add" v-if="owned">
+                <div class="label-main" @click="openLabelManager">+</div>
+            </div>
+        </div>
+        <div class="spanrow header" v-t="'common.annotation'"></div>
+        <textarea class="spanrow" :value="shape.annotation" @change="updateAnnotation" :disabled="!owned"></textarea>
+    </div>
+</template>
+
+<style scoped>
+.panel {
+    grid-template-columns: [name] 1fr [edit] 30px [move] 30px [vision] 30px [remove] 30px [end];
+    grid-column-gap: 5px;
+    align-items: center;
+    padding-bottom: 1em;
+}
+
+.label {
+    display: inline-flex;
+    position: relative;
+    flex-direction: row;
+    align-items: center;
+    background-color: white;
+    font-size: 13px;
+    margin: 5px;
+}
+
+textarea {
+    min-width: 15vh;
+    min-height: 5vh;
+}
+
+.label:hover > .label-main::before {
+    content: "\00D7";
+    position: absolute;
+    color: red;
+    font-size: 20px;
+    font-weight: bold;
+    cursor: pointer;
+    top: -8px;
+    right: -4px;
+    pointer-events: auto;
+}
+
+#label-add:hover > .label-main {
+    pointer-events: auto;
+    cursor: pointer;
+    color: white;
+    font-weight: bold;
+    background-color: #ff7052;
+}
+
+#label-add:hover > .label-main::before {
+    content: "";
+}
+
+.label-user {
+    border-top-left-radius: 10px;
+    border-bottom-left-radius: 10px;
+    background-color: #ff7052;
+    border: solid 1px #ff7052;
+    padding: 5px;
+}
+
+.label-main {
+    border: solid 1px #ff7052;
+    border-radius: 10px;
+    padding: 5px;
+    pointer-events: none;
+}
+
+.label-user + .label-main {
+    border-top-left-radius: 0;
+    border-bottom-left-radius: 0;
+}
+</style>
